correcting code to match folder refactor - main_test.py now runs
